refactor: reorganize component imports and update styling in Workflow and SidePanel
