+++ client/src/components/dashboard/workspace/sidepanel/NodeCard.tsx
