+++ client/src/components/dashboard/workspace/workflow/customNodes/TriggerNode.tsx
